[fix/#151] 공통 데이터 클래스 수정
