+++ app/src/main/java/com/bongtu/baekseo/data/dto/auth/PostTokenReissueResponse.kt
@@ -9,14 +9,4 @@ data class PostTokenReissueResponse(
     val accessToken: TokenInfo,
     @SerialName("refreshToken")
     val refreshToken: TokenInfo,
-){
-    @Serializable
-    data class TokenInfo(
-        @SerialName("token")
-        val token: String,
-        @SerialName("expiredAt")
-        val expiredAt: Long,
-        @SerialName("calculatedExpiredAt")
-        val calculatedExpiredAt: Long,
-    )
-}
+)
